Show local path instead of URI where possible
Not every document provider includes a path in the URI, which makes it
impossible to distinguish between multiple devices. Since mass storage
device emulation can only work with local files, we can try to show the
local path instead. If the file is no longer accessible (eg. due to it
being deleted), the UI will fall back to the current behavior of showing
the URI.

Fixes: #28

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>